Make changes based on Vaclav's feedback.
