+++ SplunkAppStart/Sources/SplunkAppStart/Destination/DebugDestination.swift
@@ -22,6 +22,8 @@ import SplunkSharedProtocols
 /// Stores and prints results.
 class DebugDestination: AppStartDestination {
 
+    // MARK: - Private
+
     // Stored results
     var type: AppStartType?
     var startTime: Date?
@@ -29,6 +31,9 @@ class DebugDestination: AppStartDestination {
 
     // Internal Logger
     let internalLogger = InternalLogger(configuration: .default(subsystem: "Splunk Agent", category: "AppStart"))
+
+
+    // MARK: - Sending
 
     func send(type: AppStartType, start: Date, end: Date, sharedState: (any AgentSharedState)?) {
         self.type = type
